LWR-83: adjust/clean from review
