--- src/stories/Molecules/AssetBundle/AssetBundle.css
@@ -1,18 +1,0 @@
-.asset-bundle {
-  padding: var(--s1);
-}
-
-.asset-bundle__body,
-.asset-bundle__primary-files,
-.asset-bundle__translated-files {
-  margin-bottom: var(--s1);
-}
-
-.asset-bundle__details {
-  margin-bottom: var(--s4);
-}
-
-.asset-bundle__doctype {
-  text-transform: uppercase;
-}
-
