+++ src/stories/Molecules/AssetBundle/AssetBundle.twig
@@ -2,7 +2,9 @@
 {# Include composable layout classes as needed. (stack, grid, center, etc.) #}
 {# Include variant classes form above. #}
 {% set classes = [
-  'asset-bundle'
+  'asset-bundle',
+  'box',
+  'stack',
 ] %}
 
 <div {{ attributes.addClass(classes) }}>
@@ -17,7 +19,7 @@
   {% if assets %}
     <div class="asset-bundle__primary-files">
       {% for asset in assets %}
-        <a class="button button--secondary button--normal" href="{{ asset.url }}">Download <span class="asset-bundle__doctype">{{ asset.type }}</span></a>
+        <a class="button button--secondary button--normal" href="{{ asset.url }}">Download <span class="asset-bundle__doctype">{{ asset.type|upper }}</span></a>
       {% endfor %}
     </div>
   {% endif %}
