+++ src/stories/Molecules/List/List.css
@@ -1,31 +1,14 @@
-.list {}
-
-.list__container {
-  padding: var(--s1);
-}
-
-.list--no-side-margins .list__container {
-  padding: var(--s1) 0;
-}
-
-.list__item {
-  margin-left: var(--s1);
-}
-
 .list__item-label {
-  margin-right: var(--s-6);
+  margin-inline-end: var(--s-6);
   font-weight: bold;
   text-transform: capitalize;
 }
 
-.list--no-marker .list__container {
+.list--none .list__container {
   list-style: none;
 }
 
-.list--no-marker .list__item {
+.list--none .list__item,
+.grid .list__item {
   margin-left: 0;
-}
-
-grid.list__container {
-  display: grid;
 }
