--- src/stories/Molecules/List/List.js
@@ -1,10 +1,0 @@
-/**
- * List behavior.
- *
- * DELETE THIS FILE IF IT'S NOT NEEDED
- * REMEMBER TO ALSO REMOVE THE import FROM List.stories.js
- *
- * Remember jQuery is deprecated. Go vanilla!
- * This will be wrapped in Drupal.behaviors on build, so no need for that here.
- * Just write plain old Javascript.
- **/
