--- src/stories/Organisms/Asset/Asset.js
@@ -1,10 +1,0 @@
-/**
- * Asset behavior.
- *
- * DELETE THIS FILE IF IT'S NOT NEEDED
- * REMEMBER TO ALSO REMOVE THE import FROM asset.stories.js
- *
- * Remember jQuery is deprecated. Go vanilla!
- * This will be wrapped in Drupal.behaviors on build, so no need for that here.
- * Just write plain old Javascript.
- **/
