--- src/stories/Organisms/Asset/Asset.twig
@@ -1,1 +1,0 @@
-{% include "@organisms/Section/Section.twig" with _context %}
