+++ src/stories/Templates/Asset/Asset.twig
@@ -1,0 +1,1 @@
+{% include "@organisms/Section/Section.twig" with _context %}
